Support multiple Virtual Links per NF-FG
Previsouly, the Virtual Links were created with a name that only
depended on the two LSIs they were connecting. Hence, it was not
possible to have multiple Virtual Links between the same pair of
LSIs, which is needed when they are used to connect different ports
in either of the LSIs.